+++ pephub/crud.py
@@ -1,5 +1,5 @@
 from pepdbagent import Connection
 
 
-def get_pep(db: Connection, namespace: str, pep_id: str):
-    return db.get_project(f"{namespace}/{pep_id}")
+def get_pep(db: Connection, namespace: str, project: str):
+    return db.get_project(f"{namespace}/{project}")
