+++ pephub/main.py
@@ -10,7 +10,16 @@ from starlette.staticfiles import StaticFiles
 from ._version import __version__ as server_v
 from .const import LOG_FORMAT, PKG_NAME, TAGS_METADATA
 from .helpers import build_parser
-from .routers import index_page, auth, namespace, project, eido, pep, profile
+from .routers.api.v1.base import api as api_base
+from .routers.api.v1.namespace import namespace as api_namespace
+from .routers.api.v1.project import project as api_project
+from .routers.api.v1.user import user as api_user
+from .routers.auth.base import auth as auth_router
+from .routers.views.base import views as views_base
+from .routers.views.namespace import namespace as views_namespace
+from .routers.views.project import project as views_project
+from .routers.views.submit import submit as views_submit
+from .routers.views.user import user as views_user
 from .const import STATICS_PATH, EIDO_PATH
 
 # build server
@@ -32,13 +41,16 @@ app.add_middleware(
 )
 
 # build routes
-app.include_router(index_page.router)
-app.include_router(auth.router)
-app.include_router(namespace.router)
-app.include_router(project.router)
-app.include_router(eido.router)
-app.include_router(pep.router)
-app.include_router(profile.router)
+app.include_router(api_base)
+app.include_router(api_namespace)
+app.include_router(api_project)
+app.include_router(api_user)
+app.include_router(auth_router)
+app.include_router(views_base)
+app.include_router(views_namespace)
+app.include_router(views_project)
+app.include_router(views_submit)
+app.include_router(views_user)
 
 # mount the landing html/assets
 app.mount("/static", StaticFiles(directory=STATICS_PATH), name="root_static")
