+++ pephub/route_examples.py
@@ -6,7 +6,7 @@ example_namespace = Path(
     ..., description="A namespace that holds projects.", regex=r"^\w+$", example="demo"
 )
 
-example_pep_id = Path(
+example_project = Path(
     ...,
     description="A project name inside a particular namespace",
     example="BiocProject",
