+++ pephub/routers/api/v2/__init__.py
@@ -1,0 +1,1 @@
+# future api versions will be here
