+++ pephub/templates/base/profile.html
@@ -3,12 +3,12 @@
   <div class="container">
     <div class="w-100 pb-5 border-secondary border-bottom">
         <div class="d-flex flex-row align-items-start justify-content-between mb-3">
-          <div class="d-flex flex-row align-items-center">
+          <div class="d-flex flex-row align-items-center w-75">
           <img src="{{ session_info['avatar_url'] }}" alt="Avatar" class="rounded-circle me-3 border border-secondary" height="125">
-          <div>
+          <div class="w-100">
             <h1 class="fw-bold mb-0">{{ session_info['name'] }}</h1>
             <p class="h5 fw-base mb-1">{{ session_info['login'] }}</p>
-            <p>{{ session_info['bio'] }}</p>
+            <p class=>{{ session_info['bio'] }}</p>
           </div>
           </div>
           <div>
