+++ pephub/templates/base/project.html
@@ -1,7 +1,6 @@
 {% extends "base.html" %}
 {% block content %}
 <div>
-    <a href="/pep/{{namespace}}/view">← back to {{ namespace }}</a>
     <div class="d-flex flex-row align-items-start justify-content-between">
       <div class="d-flex flex-row align-items-center">
         <h1>
@@ -10,11 +9,15 @@
         <span class="mx-2 fs-5 badge rounded-pill bg-dark">{{tag}}</span>
       </div>
       <div>
-        <button class="btn btn-outline-primary">
+        <button class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#apiEndpoints">
+          <i class="bi bi-hdd-rack-fill"></i>
+          API Endpoints
+        </button>
+        <button class="btn btn-primary">
           <i class="bi bi-pencil"></i>
           Edit
         </button>
-        <button class="btn btn-outline-danger">
+        <button class="btn btn-danger">
           <i class="bi bi-trash3"></i>
           Delete
         </button>
@@ -26,41 +29,6 @@
     <p class="mb-0">
       <span class="fw-bold">Total Samples:</span> {{ n_samples }}
     </p>
-    <div class="border-bottom border-secondary my-2"></div>
-    <h3>API Endpoints:</h3>
-      <p class="mb-2">
-        <span class="badge bg-primary">GET</span>
-        <span class="fw-bold">Project:</span>
-        <a href="/pep/{{ namespace }}/{{ project.name }}"><code>/pep/{{ namespace }}/{{ project.name }}</code></a>
-     </p>
-     <p class="mb-2">
-        <span class="badge bg-primary">GET</span>
-        <span class="fw-bold">Samples:</span>
-        <a href="/pep/{{ namespace }}/{{ project.name }}/samples"><code>/pep/{{ namespace }}/{{ project.name }}/samples</code></a>
-      </p>
-      <p class="mb-2 sample-endpoint">
-        <span class="badge bg-primary">GET</span>
-        <span class="fw-bold">Specific Sample:</span>
-        <code>
-          <a class="m-0" href="/pep/{{ namespace }}/{{ project.name }}/samples/{{samples[0]['sample_name']}}" id="sample-name-link">/pep/{{ namespace }}/{{ project.name }}/samples/</a><select class="text-primary" name="Sample name" id="sample-name" onchange="toggleSampleName(this)">
-            {% for s in samples %}
-              <option>{{s['sample_name']}}</option>
-            {% endfor %}
-          </select>
-        </code>
-      </p>
-      <p class="mb-2">
-        <span class="badge bg-primary">GET</span>
-        <span class="fw-bold">Subsamples:</span>
-        <a href="/pep/{{ namespace }}/{{ project.name }}/subsamples"><code>/pep/{{ namespace }}/{{ project.name }}/subsamples</code></a>
-      </p>
-      {% for f in filters %}
-      <p class="mb-2">
-        <span class="badge bg-primary">GET</span>
-        <span class="fw-bold">Convert project to {{f}}</span>
-        <a href="/pep/{{ namespace }}/{{ project.name }}/convert?filter={{f}}"><code>/pep/{{ namespace }}/{{ project.name }}/convert?filter={{f}}</code></a>
-      </p>
-      {% endfor %}
     <div class="border-bottom border-secondary my-2"></div>
     <div class="my-2 overflow-auto">
             <table class="table table-sm table-striped" id="project-table">
@@ -86,9 +54,52 @@
                 </tbody>
             </table>
       </div>
-      <div class="my-2">
-
+      <!-- API Endpoints Modal -->
+    <div class="modal" id="apiEndpoints" tabindex="-1" aria-labelledby="apiEndpointsModal" aria-hidden="true">
+      <div class="modal-dialog modal-xl modal-dialog-centered">
+        <div class="modal-content">
+          <div class="modal-header">
+            <h1 class="modal-title fs-5" id="apiEndpointsModal">API Endpoints Available</h1>
+            <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+          </div>
+          <div class="modal-body">
+            <p class="mb-2">
+              <span class="badge bg-primary">GET</span>
+              <span class="fw-bold">Project:</span>
+              <a href="/pep/{{ namespace }}/{{ project.name }}"><code>/pep/{{ namespace }}/{{ project.name }}</code></a>
+           </p>
+           <p class="mb-2">
+              <span class="badge bg-primary">GET</span>
+              <span class="fw-bold">Samples:</span>
+              <a href="/pep/{{ namespace }}/{{ project.name }}/samples"><code>/pep/{{ namespace }}/{{ project.name }}/samples</code></a>
+            </p>
+            <p class="mb-2 sample-endpoint">
+              <span class="badge bg-primary">GET</span>
+              <span class="fw-bold">Specific Sample:</span>
+              <code>
+                <a class="m-0" href="/pep/{{ namespace }}/{{ project.name }}/samples/{{samples[0]['sample_name']}}" id="sample-name-link">/pep/{{ namespace }}/{{ project.name }}/samples/</a><select class="text-primary" name="Sample name" id="sample-name" onchange="toggleSampleName(this)">
+                  {% for s in samples %}
+                    <option>{{s['sample_name']}}</option>
+                  {% endfor %}
+                </select>
+              </code>
+            </p>
+            <p class="mb-2">
+              <span class="badge bg-primary">GET</span>
+              <span class="fw-bold">Subsamples:</span>
+              <a href="/pep/{{ namespace }}/{{ project.name }}/subsamples"><code>/pep/{{ namespace }}/{{ project.name }}/subsamples</code></a>
+            </p>
+            {% for f in filters %}
+            <p class="mb-2">
+              <span class="badge bg-primary">GET</span>
+              <span class="fw-bold">Convert project to {{f}}</span>
+              <a href="/pep/{{ namespace }}/{{ project.name }}/convert?filter={{f}}"><code>/pep/{{ namespace }}/{{ project.name }}/convert?filter={{f}}</code></a>
+            </p>
+            {% endfor %}
+          </div>
+        </div>
       </div>
+    </div>
     </div>
   <!-- DataTabes-->
   <script src="https://cdn.jsdelivr.net/npm/simple-datatables@latest"></script>
